Adding Complete Flow for Pharmacy, Supplier and Manufacturer along with Medicine inventory management

+++ Lab10/Lab10/db4odemo/src/Business/Order/OrderDirectory.java
@@ -28,7 +28,7 @@ public class OrderDirectory {
     {
         OrderList= new ArrayList<>();
     }
-     public Order AddOrder(String medname,int Price, int units){
+     public Order AddOrder(String medname,float Price, int units){
         Order o = new Order();
         o.setItem(new Item());
         o.getItem().setProduct_name(medname);
